Returns a segment of character strings

str_mid() returns a specific number of characters from a text string, starting at the position you specify, based on the number of characters you specify.

Usage

str_mid(string, start, n)

Arguments

string

The text string containing the characters you want to extract.

start

The position of the first character you want to extract in the text. The first character in text has start = 1, and so on.

n

The length of character to extract.

Value

A character vector.

See also

Examples

str_mid("Super Eagle", 7, 5)
#> [1] "Eagle"

str_mid("Oyo Ibadan", 5, 6)
#> [1] "Ibadan"